+++ src/ReaderEither.ts
@@ -33,13 +33,14 @@ import {
   fromOptionK as fromOptionK_,
   fromPredicate as fromPredicate_
 } from './FromEither'
-import { flow, identity, pipe, Predicate, Refinement } from './function'
+import { flow, identity, pipe, Refinement } from './function'
 import { bindTo as bindTo_, flap as flap_, Functor3 } from './Functor'
 import * as _ from './internal'
 import { Monad3, Monad3C } from './Monad'
 import { MonadThrow3, MonadThrow3C } from './MonadThrow'
 import { Monoid } from './Monoid'
 import { Pointed3 } from './Pointed'
+import { Predicate } from './Predicate'
 import * as R from './Reader'
 import { Semigroup } from './Semigroup'