+++ src/ReadonlySet.ts
@@ -2,13 +2,14 @@
  * @since 2.5.0
  */
 import { Either } from './Either'
-import { Monoid } from './Monoid'
-import { Ord } from './Ord'
-import { Semigroup } from './Semigroup'
 import { Eq, fromEquals } from './Eq'
-import { Predicate, not, Refinement, identity } from './function'
-import { Separated, separated } from './Separated'
+import { identity, Refinement } from './function'
+import { Monoid } from './Monoid'
 import { Option } from './Option'
+import { Ord } from './Ord'
+import { not, Predicate } from './Predicate'
+import { Semigroup } from './Semigroup'
+import { Separated, separated } from './Separated'
 import { Show } from './Show'
 
 /**