+++ src/index.ts
@@ -69,6 +69,7 @@ import * as ord from './Ord'
 import * as ordering from './Ordering'
 import * as pipeable from './pipeable'
 import * as pointed from './Pointed'
+import * as predicate from './Predicate'
 import * as profunctor from './Profunctor'
 import * as random from './Random'
 import * as reader from './Reader'
@@ -378,6 +379,10 @@ export {
    * @since 2.10.0
    */
   pointed,
+  /**
+   * @since 2.11.0
+   */
+  predicate,
   /**
    * @since 2.0.0
    */